refactor: solve picker circular dependency issue
