+++ src/panels/DecadePanel/constant.ts
@@ -1,0 +1,2 @@
+export const DECADE_UNIT_DIFF = 10;
+export const DECADE_DISTANCE_COUNT = DECADE_UNIT_DIFF * 10;
