+++ src/panels/YearPanel/constant.ts
@@ -1,0 +1,1 @@
+export const YEAR_DECADE_COUNT = 10;
