+++ src/utils/dateUtil.ts
@@ -1,6 +1,6 @@
-import { DECADE_UNIT_DIFF } from '../panels/DecadePanel/index';
 import type { PanelMode, NullableDateType, PickerMode, Locale, CustomFormat } from '../interface';
 import type { GenerateConfig } from '../generate';
+import { DECADE_UNIT_DIFF } from '../panels/DecadePanel/constant';
 
 export const WEEK_DAY_COUNT = 7;
